refactor(embit): remove additional embit_lib.cpp file
