--- modules/embit/embit_lib/embit_lib.h
@@ -1,8 +1,0 @@
-#include <cstddef>
-#include <cstdint>
-
-extern "C" {
-    bool embit_miniscript_parse(const char* input);
-    bool embit_descriptor_parse(const char* input);
-    char* embit_psbt_parse(const uint8_t* data, size_t len);
-}
